+++ src/main/java/team/cheese/service/SaleService.java
@@ -2,6 +2,7 @@ package team.cheese.service;
 
 import team.cheese.domain.SaleDto;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface SaleService {
@@ -9,7 +10,7 @@ public interface SaleService {
     int getCount() throws Exception;
 
     // 판매자가 자신의 게시글을 삭제할 때
-    int remove(Integer no, String writer) throws Exception;
+    int remove(BigInteger no, String writer) throws Exception;
 
     // 판매자가 판매 게시글을 작성할 때
     int write(SaleDto saleDto) throws Exception;
@@ -18,7 +19,7 @@ public interface SaleService {
     List<SaleDto> getList() throws Exception;
 
     // 판매글 하나에 들어가서 게시글을 읽을 때
-    SaleDto read(Integer no) throws Exception;
+    SaleDto read(BigInteger no) throws Exception;
 
     // 판매 게시글을 수정할 때
     int modify(SaleDto saleDto) throws Exception;
